requirements.txt
asgiref==3.8.1
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-markdownify==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
language-tool-python==2.8
Markdown==3.6
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tinycss2==1.2.1
tqdm==4.66.5
urllib3==2.2.2
webencodings==0.5.1
wheel==0.44.0
